Feat: Se arregla código hijo y código padre en términos del Buffer de salida y entrada.

+++ hijo/child.c
@@ -7,19 +7,19 @@
 #include "../include/child.h"
 
 // Procedimiento que dado la sumatoria de valores reales de la visibilidad y el número de visibilidad, calcula la media real.
-// Entrada: double que representa la sumatoria de valores reales de la visibilidad y el número de visibilidad.
-// Salida: double que contiene el resultado de la media real.
-double realAverage(double sumR, double n){
+// Entrada: float que representa la sumatoria de valores reales de la visibilidad y el número de visibilidad.
+// Salida: float que contiene el resultado de la media real.
+float realAverage(float sumR, float n){
 
-    double cte = 1/n;
+    float cte = 1/n;
     return cte * sumR;
 
 }
 
 // Procedimiento que dado la sumatoria de valores imaginarios de la visibilidad y el número de visibilidad, calcula la media real.
-// Entrada: double que representa la sumatoria de valores imaginarios de la visibilidad y el número de visibilidad.
-// Salida: double que contiene el resultado de la media imaginaria.
-double imaginaryAverage(double sumI, double n){
+// Entrada: float que representa la sumatoria de valores imaginarios de la visibilidad y el número de visibilidad.
+// Salida: float que contiene el resultado de la media imaginaria.
+float imaginaryAverage(float sumI, float n){
 
     int cte = 1/n;
     return cte * sumI;
@@ -27,13 +27,13 @@ double imaginaryAverage(double sumI, double n){
 }
 
 // Procedimiento que dado la sumatoria de valores imaginarios y la sumatoria de valores reales de la visibilidad, entrega la potencia de la muestra de visibilidades.
-// Entrada: double que representa la sumatoria de valores reales de la visibilidad e double que representa la sumatoria de valores imaginarios.
-// Salida: double que contiene el resultado de la potencia de la muestra de visibilidades.
-double potency(double sumR, double sumI){
+// Entrada: float que representa la sumatoria de valores reales de la visibilidad e float que representa la sumatoria de valores imaginarios.
+// Salida: float que contiene el resultado de la potencia de la muestra de visibilidades.
+float potency(float sumR, float sumI){
 
-    double sumR2 = pow(sumR,2);
-    double sumI2 = pow(sumI,2);
-    double sum = sumR2 + sumI2;
+    float sumR2 = pow(sumR,2);
+    float sumI2 = pow(sumI,2);
+    float sum = sumR2 + sumI2;
     return sqrt(sum);
 
 }
